+++ packages/editor/src/blocks/endNode.ts
@@ -4,6 +4,7 @@ import { NodeData } from "@/types/NodeTypes";
 export const endNode: Partial<Node<NodeData>> = {
   type: "custom",
   data: {
+    addable: false,
     block: "end@0.1",
     label: "End",
     color: "rgb(239 68 68)",
