+++ packages/editor/src/blocks/logicNode.ts
