+++ packages/editor/src/blocks/startNode.ts
@@ -4,6 +4,7 @@ import { Node } from "reactflow";
 export const startNode: Partial<Node<NodeData>> = {
   type: "custom",
   data: {
+    addable: false,
     block: "start@0.1",
     label: "Start",
     color: "rgb(34 197 94)",
