+++ packages/editor/src/components/controls/AddBlockDropdown.tsx
@@ -1,15 +1,10 @@
-import {
-  TextCursorInput,
-  Square,
-  GitBranch,
-  Plus,
-  Play,
-  Flag,
-} from "lucide-react";
+import React from "react";
+import { Plus } from "lucide-react";
 import { useDispatch } from "react-redux";
 import { startDraggingNode } from "../../slices/reactFlowSlice";
-import { NodeType } from "../../nodes";
+import { blockConfigs, BlockType } from "../../blocks";
 import { useReactFlow } from "reactflow";
+import NodeIcon from "../NodeIcon";
 
 import {
   DropdownMenu,
@@ -18,9 +13,6 @@ import {
   DropdownMenuItem,
   DropdownMenuLabel,
   DropdownMenuSeparator,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from "@data-river/shared/ui/components/ui/dropdown-menu";
 
@@ -28,14 +20,6 @@ interface AddBlockDropdownMenuProps {
   direction?: "up" | "down";
   children: React.ReactNode;
 }
-
-const nodeTypeIcons = {
-  start: Play,
-  input: TextCursorInput,
-  logic: GitBranch,
-  output: Square,
-  end: Flag,
-};
 
 export function AddBlockDropdownMenu({
   direction = "down",
@@ -46,7 +30,7 @@ export function AddBlockDropdownMenu({
 
   const handleStartDraggingNode = (
     event: React.MouseEvent<HTMLDivElement>,
-    type: NodeType,
+    type: BlockType,
   ) => {
     const position = screenToFlowPosition({
       x: event.clientX,
@@ -54,6 +38,25 @@ export function AddBlockDropdownMenu({
     });
 
     dispatch(startDraggingNode({ type, position }));
+  };
+
+  const renderMenuItems = () => {
+    return Object.entries(blockConfigs)
+      .filter(([_, config]) => config.data?.addable ?? true)
+      .map(([type, config]) => (
+        <DropdownMenuItem
+          key={type}
+          onClick={(event) => handleStartDraggingNode(event, type as BlockType)}
+        >
+          <NodeIcon
+            icon={config.data!.icon}
+            color={config.data!.color}
+            useBackgroundColor={false}
+            useIconColor={false}
+          />
+          <span className="ml-2">{config.data!.label}</span>
+        </DropdownMenuItem>
+      ));
   };
 
   return (
@@ -68,42 +71,9 @@ export function AddBlockDropdownMenu({
       >
         <DropdownMenuLabel>Add Block to Flow</DropdownMenuLabel>
         <DropdownMenuSeparator />
-        <DropdownMenuGroup>
-          <DropdownMenuSub>
-            <DropdownMenuSubTrigger>
-              <TextCursorInput className="mr-2 h-4 w-4" />
-              <span>Input</span>
-            </DropdownMenuSubTrigger>
-            <DropdownMenuSubContent>
-              <DropdownMenuItem
-                onClick={(event) => handleStartDraggingNode(event, "input")}
-              >
-                <TextCursorInput className="mr-2 h-4 w-4" />
-                <span>Input Node</span>
-              </DropdownMenuItem>
-              <DropdownMenuItem
-                onClick={(event) => handleStartDraggingNode(event, "input")}
-              >
-                <Square className="mr-2 h-4 w-4" />
-                <span>Simple Input</span>
-              </DropdownMenuItem>
-            </DropdownMenuSubContent>
-          </DropdownMenuSub>
-          <DropdownMenuItem
-            onClick={(event) => handleStartDraggingNode(event, "logic")}
-          >
-            <GitBranch className="mr-2 h-4 w-4" />
-            <span>Logic</span>
-          </DropdownMenuItem>
-          <DropdownMenuItem
-            onClick={(event) => handleStartDraggingNode(event, "output")}
-          >
-            <Square className="mr-2 h-4 w-4" />
-            <span>Output</span>
-          </DropdownMenuItem>
-        </DropdownMenuGroup>
+        <DropdownMenuGroup>{renderMenuItems()}</DropdownMenuGroup>
         <DropdownMenuSeparator />
-        <DropdownMenuItem>
+        <DropdownMenuItem disabled>
           <Plus className="mr-2 h-4 w-4" />
           <span>Add Custom Block</span>
         </DropdownMenuItem>
